# docs/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "EvaluationCase": {
        "additionalProperties": false,
        "properties": {
          "cohort": {
            "default": "default",
            "maxLength": 64,
            "minLength": 1,
            "title": "Cohort",
            "type": "string"
          },
          "comparison_id": {
            "format": "uuid",
            "title": "Comparison Id",
            "type": "string"
          },
          "target_at": {
            "format": "date-time",
            "title": "Target At",
            "type": "string"
          },
          "target_item_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]*$",
            "title": "Target Item Id",
            "type": "string"
          }
        },
        "required": [
          "comparison_id",
          "target_item_id",
          "target_at"
        ],
        "title": "EvaluationCase",
        "type": "object"
      },
      "EvaluationInput": {
        "additionalProperties": false,
        "properties": {
          "cases": {
            "items": {
              "$ref": "#/components/schemas/EvaluationCase"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Cases",
            "type": "array"
          },
          "dataset_name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Dataset Name",
            "type": "string"
          },
          "k": {
            "default": 20,
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "K",
            "type": "integer"
          },
          "label_origin": {
            "enum": [
              "synthetic",
              "manual",
              "historical_export"
            ],
            "title": "Label Origin",
            "type": "string"
          },
          "schema_version": {
            "const": "0.1",
            "default": "0.1",
            "title": "Schema Version",
            "type": "string"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          },
          "source_description": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Source Description",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "dataset_name",
          "label_origin",
          "source_description",
          "cases"
        ],
        "title": "EvaluationInput",
        "type": "object"
      },
      "EvaluationJobResponse": {
        "properties": {
          "attempts": {
            "title": "Attempts",
            "type": "integer"
          },
          "cancel_requested": {
            "title": "Cancel Requested",
            "type": "boolean"
          },
          "completed_cases": {
            "title": "Completed Cases",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "dataset_name": {
            "title": "Dataset Name",
            "type": "string"
          },
          "dataset_sha256": {
            "title": "Dataset Sha256",
            "type": "string"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "label_origin": {
            "enum": [
              "synthetic",
              "manual",
              "historical_export"
            ],
            "title": "Label Origin",
            "type": "string"
          },
          "protocol": {
            "const": "held-out-saved-snapshot-v1",
            "title": "Protocol",
            "type": "string"
          },
          "replay_of": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Replay Of"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "cancelling",
              "completed",
              "cancelled",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "total_cases": {
            "title": "Total Cases",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "session_id",
          "status",
          "completed_cases",
          "total_cases",
          "attempts",
          "cancel_requested",
          "error_code",
          "replay_of",
          "dataset_name",
          "label_origin",
          "dataset_sha256",
          "protocol",
          "created_at",
          "updated_at"
        ],
        "title": "EvaluationJobResponse",
        "type": "object"
      },
      "EvaluationPageResponse": {
        "properties": {
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/EvaluationJobResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "offset",
          "limit",
          "has_more"
        ],
        "title": "EvaluationPageResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "本地推荐与管理演示。未配置 PostgreSQL：会话和推荐结果只保存在当前进程。受控 CPU bundle 可发布；R06 冻结语料服务须显式启用并批准发布，不支持任意新商品。",
    "title": "EvoRec Demo API",
    "version": "0.0.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/evaluation-jobs": {
      "get": {
        "operationId": "listing_api_v1_evaluation_jobs_get",
        "parameters": [
          {
            "in": "query",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 10000,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Session-Token",
            "required": true,
            "schema": {
              "title": "X-Session-Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Listing",
        "tags": [
          "evaluation"
        ]
      },
      "post": {
        "operationId": "enqueue_api_v1_evaluation_jobs_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Session-Token",
            "required": true,
            "schema": {
              "title": "X-Session-Token",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluationInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enqueue",
        "tags": [
          "evaluation"
        ]
      }
    },
    "/api/v1/evaluation-jobs/{job_id}": {
      "get": {
        "operationId": "get_api_v1_evaluation_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Session-Token",
            "required": true,
            "schema": {
              "title": "X-Session-Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get",
        "tags": [
          "evaluation"
        ]
      }
    },
    "/api/v1/evaluation-jobs/{job_id}/cancel": {
      "post": {
        "operationId": "cancel_api_v1_evaluation_jobs__job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Session-Token",
            "required": true,
            "schema": {
              "title": "X-Session-Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel",
        "tags": [
          "evaluation"
        ]
      }
    },
    "/api/v1/evaluation-jobs/{job_id}/replay": {
      "post": {
        "operationId": "replay_api_v1_evaluation_jobs__job_id__replay_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Session-Token",
            "required": true,
            "schema": {
              "title": "X-Session-Token",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replay",
        "tags": [
          "evaluation"
        ]
      }
    },
    "/api/v1/evaluation-jobs/{job_id}/report": {
      "get": {
        "operationId": "report_api_v1_evaluation_jobs__job_id__report_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "download",
            "required": false,
            "schema": {
              "default": false,
              "title": "Download",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "X-Session-Token",
            "required": true,
            "schema": {
              "title": "X-Session-Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Report",
        "tags": [
          "evaluation"
        ]
      }
    },
...
  }
}